Working version
Node class is now in its own file. Board class has been completely changed, create_board method has been replaced by create_children. Instead of building the board and then trying to traverse it, the squares are built on the fly as children of the square the knight is on.